==> 00-prep_GeoNames-preparation.ipynb <==
# GeoNames preparation
Optionally download GeoNames data for Switzerland, then project records to swiss coordinates (E and N) and save to file.

import pandas as pd
import time
import os

# own modules
from gazmatch import gazetteers
from gazmatch import projection

data_dir = 'data'

### Optional: download latest GeoNames Switzerland data

import urllib.request
from zipfile import ZipFile

geonames_ch_zip_url = r'http://download.geonames.org/export/dump/CH.zip'
geonames_ch_zip_local = os.path.join(data_dir, 'CH.zip')

urllib.request.urlretrieve(geonames_ch_zip_url, geonames_ch_zip_local)
with ZipFile(geonames_ch_zip_local, 'r') as zipfile:
    zipfile.extract('CH.txt', data_dir)

os.remove(geonames_ch_zip_local)

### Load GeoNames

geonames = gazetteers.GeoNamesCH(data_dir=data_dir, verbose=True)
geonames.df.head()

geonames_copy = geonames.df.copy()
print(geonames_copy.shape)

### Project all features in batches

max_features = 20000
start_index = 0
dfs = []
while (start_index+1) <= geonames_copy.shape[0]:
    print("Start index was %s" %start_index)
    geonames_subset = geonames_copy.iloc[start_index:start_index+max_features].copy() # select feature subset
    
    lats = geonames_subset['latitude'].tolist()
    lons = geonames_subset['longitude'].tolist()
    latlon_coords = list(zip(lats, lons))
    print("Started with %s lat-lon coordinates." %len(latlon_coords))
    t1 = time.time()
    swiss_coords = projection.latlon_to_swiss_batch(latlon_coords)
    t2 = time.time()
    print("Ended up with %s swiss coordinates." %len(swiss_coords))
    print('Projection took %0.3fs' % ((t2-t1)))
    Es, Ns = zip(*swiss_coords)
    geonames_subset['gn_E'] = pd.Series(list(Es), index=geonames_subset.index)
    geonames_subset['gn_N'] = pd.Series(list(Ns), index=geonames_subset.index)    
    dfs.append(geonames_subset)
    start_index += max_features

#print("\nWe end up with %s dataframes" %len(dfs))

### Concat back into one

df_all = pd.concat(dfs, axis=0)
df_all.shape

# sanity check (there should not be any nulls)
print(df_all[df_all['gn_E'].isnull()].shape)
print(df_all[df_all['gn_N'].isnull()].shape)

# export to csv
path_filename_csv = os.path.join(data_dir, 'geonames_ch_swisscoords.csv')
df_all.to_csv(path_or_buf=path_filename_csv, sep='\t', index=False, encoding='utf-8')
