==> 01-prep_SwissNames3D-preparation.ipynb <==
# Prepare SwissNames3D
We project all records to lat,lon and we create unique IDs for each row, rather than for each geographical feature (which may have several rows when it has multiple names). The new UUIDs will sit in a new column.

import pandas as pd
import os
import time

# own modules
from gazmatch import gazetteers
from gazmatch import projection

### Start from the original SwissNames3D data
The data can be downloaded from here: https://shop.swisstopo.admin.ch/en/products/landscape/names3D .

# path to SwissNames3D gazetteer CSV data, LV03 projection
swissnames_data_dir = "data\\SwissNames3D\\swissNAMES3D_LV03\\csv_LV03_LN02\\"
swissnames = gazetteers.SwissNames3D(data_dir=swissnames_data_dir, verbose=True)

swissnames.df.head()

### Project each record to lat lon

swissnames_copy = swissnames.df.copy()
print(swissnames_copy.shape)

max_features = 25000
start_index = 0
dfs = []
while (start_index+1) <= swissnames_copy.shape[0]:
    print("Start index was %s" %start_index)
    swissnames_subset = swissnames_copy.iloc[start_index:start_index+max_features].copy() # select 25k features
    Es = swissnames_subset['E'].tolist()
    Ns = swissnames_subset['N'].tolist()
    swiss_coords = list(zip(Es, Ns))
    time1 = time.time()
    print("Started with %s swiss coordinates." %len(swiss_coords))
    lat_lon_coords = projection.swiss_to_latlon_batch(swiss_coords)
    print("Ended up with %s lat-lon coordinates." %len(lat_lon_coords))
    time2 = time.time()
    #print('Projection took %0.3fs' % ((time2-time1)))
    lats, lons = zip(*lat_lon_coords)
    swissnames_subset['sn_lat'] = pd.Series(list(lats), index=swissnames_subset.index)
    swissnames_subset['sn_lon'] = pd.Series(list(lons), index=swissnames_subset.index)
    dfs.append(swissnames_subset)
    start_index += max_features

print(len(dfs))

# concatenate back into one
df_all = pd.concat(dfs, axis=0)
df_all.shape

# sanity check (there should not be any nulls)
print(df_all[df_all['sn_lat'].isnull()].shape)
print(df_all[df_all['sn_lon'].isnull()].shape)

### Prepare to create new UUIDs

# how many UUIDs appear more than once and how many times?
id_groups = df_all.groupby(['UUID'])
print("We have %s id groups for %s features" %(len(id_groups), df_all.shape[0]))
id_groups.size().value_counts().sort_index()

# we want the new column to be called 'UUID', so rename the current one
df_all = df_all.rename(columns={'UUID': 'UUID_old'})
df_all.columns

# current sprachcode values
df_all['SPRACHCODE'].value_counts(dropna=False)

# use sprachcodes as a second disambiguator in the new UUID, alongside the name
sprachcode_to_lang = {}
sprachcode_to_lang['Hochdeutsch inkl. Lokalsprachen'] = 'de'
sprachcode_to_lang['Franzoesisch inkl. Lokalsprachen'] = 'fr'
sprachcode_to_lang['Italienisch inkl. Lokalsprachen'] = 'it'
sprachcode_to_lang['Rumantsch Grischun inkl. Lokalsprachen'] = 'rm'
sprachcode_to_lang['Mehrsprachig'] = 'me'
sprachcode_to_lang['ub'] = 'na'
sprachcode_to_lang['nan'] = 'na'

def add_sprach_code(row):
    lang_id = sprachcode_to_lang[str(row['SPRACHCODE'])]
    return lang_id

# add a row with our shorter sprachcode strings
df_all['lang_id'] = df_all.apply(add_sprach_code, axis=1)

# sanity check
df_all['lang_id'].value_counts(dropna=False)

### Create new UUIDs

def deduplicate_uuids(row):
    name_no_spaces = str(row['NAME']).replace(' ', '')
    new_uuid = row['UUID_old'] + '-' + name_no_spaces + '-' + row['lang_id']
    return new_uuid

# make the new UUIDs for all records
df_all['UUID'] = df_all.apply(deduplicate_uuids, axis=1)

# check that our new UUIDs are all unique
id_groups = df_all.groupby(['UUID'])
print("We now have %s id groups for %s records" %(len(id_groups), df_all.shape[0]))
id_groups.size().value_counts().sort_index()

# view a sample of the new UUIDs
df_all['UUID'].sample(n=10).tolist()

### Export to CSV

df_all.columns

# put the UUID column where we want it: at the front and next to the UUID_old
col_order = ['UUID', 'UUID_old', 'NAME', 'GEOMTYPE', 'OBJEKTKLASSE_TLM', 'OBJEKTART', 'SPRACHCODE', 'lang_id', 
             'E', 'N', 'Z', 'HOEHE', 'sn_lat', 'sn_lon', 'EINWOHNERKATEGORIE', 'POPCATINT', 'GEBAEUDENUTZUNG', 
             'KUNSTBAUTE', 'NAMENGRUPPE_UUID', 'NAMEN_TYP', 'NAME_UUID']
df_all = df_all[col_order]
df_all.head()

# export to csv
output_dir = 'data'
filename_csv = 'swissnames_new_uuids.csv'
df_all.to_csv(path_or_buf=os.path.join(output_dir, filename_csv), sep='\t', index=False, encoding='utf-8')
