==> 02-prep_matching-ground-truth.ipynb <==
# Matching preparation
- Build a flat ground truth structure which is just a list of matches (rather than having extra_matches listed in a 3rd column in a table, they are all treated as just another matching pair).
- Augment the ground truth by 're-hydrating' our data, going from the geonames IDs that we annotated to a fuller table where each ID is now a row with extra record info contained in GeoNames.

import pandas as pd
import pickle
import os

# own file
from gazmatch import gazetteers

data_dir = 'data'
path_to_ground_truth = os.path.join(data_dir, 'annotated_sample.csv')
df_matches_grouped = pd.read_csv(path_to_ground_truth, sep='\t', encoding='utf-8')
df_matches_grouped.head()

### Flatten extra matches

ground_truth_flat = []

for index, row in df_matches_grouped.iterrows():
    gn_id = row['geonameid']
    sn_id = str(row['swissnamesid'])
    if sn_id.startswith("{"):
        ground_truth_flat.append((gn_id, sn_id))
    else:
        # we don't even have a top match
        continue
    sn_extra_ids = str(row['extra_matches'])
    # process the extra matches if we have any 
    if sn_extra_ids != 'nan':
        for sn_extra_id in str(sn_extra_ids).split(";"):
            sn_id_clean = sn_extra_id.strip()
            if sn_id_clean.startswith("{"):
                ground_truth_flat.append((gn_id, sn_id_clean))

print("flattened ground truth has %s pairs" %len(ground_truth_flat))

# turn into a dataframe
cols = ['geonameid', 'swissnamesid']
df_matches_flat = pd.DataFrame.from_records(ground_truth_flat, columns=cols)
df_matches_flat.head()

# export to CSV
path_to_ground_truth_flat = os.path.join(data_dir, 'ground_truth_flat.csv')
df_matches_flat.to_csv(path_to_ground_truth_flat, sep='\t', index=False, encoding='utf-8')

# also pickle (serialize) the list directly
path_to_pickle = os.path.join(data_dir, 'ground_truth_flat.pkl')
with open(path_to_pickle, 'wb') as pickle_file:
    pickle.dump(ground_truth_flat, pickle_file)

### Augment ground truth ids

geonames = gazetteers.GeoNamesCH(data_dir=data_dir, verbose=True)

geoname_ids = df_matches_grouped['geonameid'].tolist()
len(geoname_ids)

augmented_records = []

for geoname_id in geoname_ids:
    record = geonames.retrieve_by_id(geoname_id)
    augmented_records.append(record)

df_augmented = pd.DataFrame(augmented_records)
df_augmented.head()

# export to CSV
path_to_augmented_data = os.path.join(data_dir, 'source_data_augmented.csv')
df_augmented.to_csv(path_to_augmented_data, sep='\t', index=False, encoding='utf-8')
