==> 10_rules_random-baseline.ipynb <==
# Rule-based gazetteer matching: random-baseline

import pandas as pd
import os
    
# own file
from gazmatch import matchutils as mu

data_dir = 'data'

### Load data

# load ground truth data (pickled)
ground_truth_flat = mu.load_data(os.path.join(data_dir, 'ground_truth_flat.pkl'))
print("%s matching pairs" %len(ground_truth_flat))

# load SwissNames3D augmented with lat, lons
path_swissnames_csv = os.path.join(data_dir, 'swissnames_new_uuids.csv')
swissnames = pd.read_csv(path_swissnames_csv, sep='\t', encoding='utf-8', low_memory=False)
swissnames.shape

# source data (Geonames records)
path_to_source_data = os.path.join(data_dir, 'source_data_augmented.csv')
df_source = pd.read_csv(path_to_source_data, sep='\t', encoding='utf-8')

## Rule-based matching

def get_exact_matches(token, grouped_df, df_cols):
    try:
        df_ret = grouped_df.get_group(token)
        return df_ret
    except KeyError:
        return pd.DataFrame(columns=df_cols)

def single_run(source_type_groups, target_name_groups, target_cols, print_type_results=False, print_overall_results=False):
    matches_overall = []
    neg_overall = []
    if print_type_results:
        print("\tprec\trecall\tf1")
    for ftype, group in source_type_groups:
        matches_type = []
        neg_type = []
        gn_ids_type = []
        for index, gn_record in group.iterrows():
            name = gn_record['name']
            gn_id = gn_record['geonameid']
            gn_ids_type.append(gn_id)
            # 1. get any exact name matches as candidates
            df_exact_all = get_exact_matches(name, target_name_groups, target_cols)
            df_exact = df_exact_all.copy()
            if df_exact.empty:
                neg_type.append((gn_id, mu.NO_MATCH_STRING))
                neg_overall.append((gn_id, mu.NO_MATCH_STRING))
                continue
            # 2. take a random exact match as the match
            df_match = df_exact.sample() # defaults to n=1
            matches_type.append((gn_id, df_match.iloc[0]['UUID_old']))
            matches_overall.append((gn_id, df_match.iloc[0]['UUID_old']))
        # filter ground truth to ids of current feature type only
        gn_ids_type_set = set(gn_ids_type)
        ground_truth_type = [item for item in ground_truth_flat if item[0] in gn_ids_type_set]
        p_type, r_type, f1_type = mu.evaluate_results_list(matches_type, neg_type, ground_truth_type, verbose=False)
        if print_type_results:
            print("%s\t%.3f\t%.3f\t%.3f" %(ftype, p_type, r_type, f1_type))
            
    p, r, f1 = mu.evaluate_results_list(matches_overall, neg_overall, ground_truth_flat, verbose=False)
    if print_overall_results:
        print("%.3f\t%.3f\t%.3f" %(p, r, f1))
    return matches_overall, neg_overall

### Run once

# swissnames name groups
sn_name_groups = swissnames.groupby(['NAME'])

# geonames type groups
type_groups = df_source.groupby(['feature code'])

# show performance by type
test_single = single_run(type_groups, sn_name_groups, swissnames.columns, print_type_results=True, print_overall_results=False)

### Run multiple times
Because there's a randomness element, the results change slightly at each run.

# run 10 times
print("prec\trecall\tf1")
for i in range(0,10):
    single_run(type_groups, sn_name_groups, swissnames.columns, print_type_results=False, print_overall_results=True)
